added function to set variable depending on block device type

+++ lvm-luks-install.sh
@@ -6,7 +6,7 @@ You will enter a password to use when encrypting the disk, it will be saved to /
 
 EOF
 echo -n "Enter Password: "
-read CRYPT_PASSWORD
+read -rs CRYPT_PASSWORD
 if [ ${#CRYPT_PASSWORD} -lt 8 ]; then echo "Password Too Short"; exit
 else echo ${CRYPT_PASSWORD} > /tmp/lukspassword
 fi
@@ -16,26 +16,39 @@ else
 	echo "Password could not be set, please check if you have write permission to /tmp"; exit
 fi
 
-echo -n "Enter disk to use: "
+echo -n "Enter disk to use (For example /dev/sda or /dev/nvme0): "
 read DISK
-if [ ! -b ${DISK} ]; then echo "Disk you entered is not a block device... exiting.."; exit
+if [ ! -b ${DISK} ]; then echo "Enter full path to block device... exiting.."; exit
 else
 	echo "Using ${DISK} to setup encrypted disk on. IT WILL BE WIPED!!"
 fi
 
-read -p "Are you sure? ([yY]es)" -n 1 -r
+read -p "Are you sure, it will wiped? ([yY]es)" -n 1 -r
 echo
-if [[ ! $REPLY =~ ^[Yy]$ ]]; then exit; fi
+if [ ! $REPLY =~ ^[Yy]$ ]; then exit; fi
 
 read -p "Installing NixOS?" -n 1 -r
 echo
-if [[ ! $REPLY =~ ^[Yy]$ ]]; then 
+if [ ! $REPLY =~ ^[Yy]$ ]; then 
 	export DISTRIBUTION="NixOS"
 fi
 
-# This will result in /dev/nvmen0p1, if disk was /dev/nvmen0
-ROOT_PART="${DISK}p1"
-BOOT_PART="${DISK}p2"
+############## VARIABLE SETTINGS
+
+# Sets root and boot partion variables depending on nvme0 type or sda/vga
+case ${DISK#/dev/} in
+	sda | vga)
+		ROOT_PART="${DISK}1"
+		BOOT_PART="${DISK}2"
+		;;
+	nvme0)
+		ROOT_PART="${DISK}p1"
+		BOOT_PART="${DISK}p2"
+		;;
+	*)
+		echo "No valid disk type.."
+		;;
+esac
 
 SWAP_ON=true
 SWAP_SIZE=8
@@ -61,7 +74,13 @@ parted ${DISK} -- mkpart primary 512MiB 100%
 echo "SETTING UP BOOT PARTITIONS"
 parted ${DISK} -- mkpart esp fat32 1MiB 512MiB
 parted ${DISK} -- set 2 esp on
-mkfs.fat -F 32 -n boot ${BOOT_PART}
+
+if [ -b ${BOOT_PART} ]; then 
+	mkfs.fat -F 32 -n boot ${BOOT_PART}
+else
+	echo "Could not find boot block device"
+	exit
+fi
 
 # SETUP CRYPT FOR DISK ENCRYPTION
 cryptsetup -y -v --type luks2 luksFormat ${ROOT_PART} -d /tmp/lukspassword
@@ -98,7 +117,7 @@ fi
 
 mount /dev/${VG_NAME}/root /mnt
 mkdir -p /mnt/boot
-mount /dev/${BOOT_PART} /mnt/boot
+mount ${BOOT_PART} /mnt/boot
 
 for lv in ${!EXTRA_LVS[@]}; do
 	mkdir -p /mnt/${lv}
